requirements.txt
aiohappyeyeballs==2.4.0
aiohttp==3.10.5
aiosignal==1.3.1
annotated-types==0.7.0
anyio==4.4.0
attrs==24.2.0
Authlib==1.3.2
cerebras_cloud_sdk==1.3.0
certifi==2024.7.4
cffi==1.17.0
charset-normalizer==3.3.2
cryptography==43.0.0
discord.py==2.4.0
distro==1.9.0
frozenlist==1.4.1
grpcio==1.66.0
grpcio-health-checking==1.66.0
grpcio-tools==1.66.0
h11==0.14.0
httpcore==1.0.5
httpx==0.27.0
idna==3.8
multidict==6.0.5
protobuf==5.27.4
pycparser==2.22
pydantic==2.8.2
pydantic_core==2.20.1
python-dotenv==1.0.1
requests==2.32.3
setuptools==74.0.0
sniffio==1.3.1
typing_extensions==4.12.2
urllib3==2.2.2
validators==0.33.0
weaviate-client==4.7.1
yarl==1.9.4
